moved ParseXMLWorker to subdir
