+++ run.py
@@ -1,8 +1,8 @@
 from pprint import pprint
-from ngxmlzip import process
-from ngxmlzip import create
+from ngxmlzip import process_runner
+from ngxmlzip import create_runner
 from ngxmlzip.utils import OperationResult
-from ngxmlzip.colors import TextColors
+from ngxmlzip.utils import TextColors
 
 MAX_OBJECTS_IN_XML = 10
 XML_FILES_IN_ZIP = 100
@@ -13,13 +13,13 @@ CSV_FILE_2 = "csv_file_2.csv"
 
 
 if __name__ == "__main__":
-    if not create.create_dir(ZIP_DIRECTORY):
+    if not create_runner.create_dir(ZIP_DIRECTORY):
         print(f"{TextColors.FAIL}directory {ZIP_DIRECTORY} could not be created!{TextColors.ENDC}")
         exit()
 
     create_result: OperationResult = None
     try:
-        create_result = create.create_zip_files(
+        create_result = create_runner.create_zip_files(
             MAX_OBJECTS_IN_XML,
             XML_FILES_IN_ZIP,
             ZIP_FILES,
@@ -41,7 +41,7 @@ if __name__ == "__main__":
     print("================ Create results  ========================")
     pprint(create_result)
     print("================ Process result =========================")
-    process_result = process.run_multi_proc(ZIP_DIRECTORY, CSV_FILE_1, CSV_FILE_2)
+    process_result = process_runner.run_multi_proc(ZIP_DIRECTORY, CSV_FILE_1, CSV_FILE_2)
     pprint(process_result)
 
     if create_result == process_result:
